fix golint
